Refactor core into smaller modules
